LPD-86558 Rename getValidateLarFileEndpoint to getValidateLarFile
