+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/js/index.ts
@@ -3,11 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {default as formatDate} from '../revamp/js/common/utils/formatDate';
 export {default as ImportReportFDSPropsTransformer} from '../revamp/js/fds/ImportReportFDSPropsTransformer';
 export {NewExport} from '../revamp/js/pages/export/NewExport';
 export {NewImport} from '../revamp/js/pages/import/NewImport';
 export {ViewImportReportEntryDetail} from '../revamp/js/pages/import/report/ViewImportReportEntryDetail';
+export {default as formatDate} from '../revamp/js/utils/formatDate';
 
 // @ts-ignore
