+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/revamp/js/pages/import/report/ViewImportReportEntryDetail.tsx
@@ -11,7 +11,7 @@ import {openModal} from 'frontend-js-components-web';
 import {escapeHTML, fetch} from 'frontend-js-web';
 import React, {useEffect, useState} from 'react';
 
-import formatDate from '../../../common/utils/formatDate';
+import formatDate from '../../../utils/formatDate';
 
 function DetailViewDefinitionCol({
 	body,
